feat(pgpm-bundle): extract @pgpmjs/bundle package; applyBundle stays in core
Move the pure migration-bundle artifact layer out of @pgpmjs/core into a new
package @pgpmjs/bundle:
- create/io/verify/transpile/diff/reconcile/types -> @pgpmjs/bundle

applyBundle stays in @pgpmjs/core: it drives PgpmMigrate (the deploy engine),
so moving it would create a core<->bundle cycle. core/bundle/index.ts now
re-exports @pgpmjs/bundle + ./apply, so the full surface stays importable from
@pgpmjs/core unchanged.

Two pure helpers the bundle layer needs move down into @pgpmjs/ast (the pure
leaf both packages depend on), each re-exported from its old core location for
compatibility:
- hashString  (was migrate/utils/hash)
- renameInPlanContent (was refactor/rename)

Also add the standard logo + CI/license/npm badges to the @pgpmjs/ast and
@pgpmjs/bundle READMEs, and add pgpm/bundle to the CI pg-core test batch.
No behavior change.
