+++ pgpm/bundle/src/index.ts
@@ -1,5 +1,5 @@
 /**
- * pgpm migration bundles: the portable, content-addressed artifact layer.
+ * @pgpmjs/bundle: the portable, content-addressed migration artifact layer.
  *
  * A {@link MigrationBundle} is a deterministic snapshot of a pgpm module — plan,
  * control, ordered changes, and per-change/whole-bundle sha256 digests — with an
@@ -7,9 +7,14 @@
  * / `applyMigrationBundle` / `transpileMigrationBundle` cloud functions: build one
  * from a module AST, transport or transform it, verify its digests, then
  * materialize it back into a deployable module.
+ *
+ * Pure layer (no database, no deploy engine). The deployment glue — `applyBundle`,
+ * which drives `PgpmMigrate` — lives in `@pgpmjs/core`.
  */
-// Pure artifact layer moved to @pgpmjs/bundle; re-exported here so existing
-// `@pgpmjs/core` consumers keep importing the whole bundle surface unchanged.
-export * from '@pgpmjs/bundle';
-// applyBundle stays in core: it drives PgpmMigrate (the deploy engine).
-export * from './apply';
+export * from './create';
+export * from './diff';
+export * from './io';
+export * from './reconcile';
+export * from './transpile';
+export * from './types';
+export * from './verify';
