+++ pgpm/core/src/bundle/apply.ts
@@ -4,11 +4,10 @@ import { join } from 'path';
 import { PgConfig } from 'pg-env';
 
 import { PgpmScriptKind } from '@pgpmjs/ast/module/types';
+import { BundleIssue, materializeBundle, MigrationBundle, verifyBundle } from '@pgpmjs/bundle';
+
 import { PgpmMigrate, PgpmMigrateOptions } from '../migrate/client';
 import { DeployResult, VerifyResult } from '../migrate/types';
-import { materializeBundle } from './io';
-import { MigrationBundle } from './types';
-import { BundleIssue, verifyBundle } from './verify';
 
 /** A script whose SQL references something outside the allowed namespace. */
 export interface NamespaceViolation {
