+++ pgpm/core/src/migrate/utils/hash.ts
@@ -1,21 +1,18 @@
+import { hashString } from '@pgpmjs/ast';
 import { createHash } from 'crypto';
 import { readFile } from 'fs/promises';
 import { parse } from 'pgsql-parser';
 
 import { cleanTree } from '../../packaging/package';
 
+// Re-exported from @pgpmjs/ast so existing `@pgpmjs/core` consumers keep importing it here.
+export { hashString };
+
 /**
  * Generate SHA256 hash of a file's contents
  */
 export async function hashFile(filePath: string): Promise<string> {
   const content = await readFile(filePath, 'utf-8');
-  return createHash('sha256').update(content).digest('hex');
-}
-
-/**
- * Generate SHA256 hash of a string
- */
-export function hashString(content: string): string {
   return createHash('sha256').update(content).digest('hex');
 }
